[loader] New linker error for duplicate entry in BinaryCache
Make it an error to try and insert the same key into a binary cache
multiple times.  Previously the cache would let the last write win,
but would leak previous entries in that case.

No existing use of `BinaryCache` is at risk of triggering this
condition, but the plan is to use a `BinaryCache` to store the global
`structs` table for canonicalization, and for that, it becomes helpful
to have this extra safety net.

Test Plan:

Existing tests:

```
$ cargo nextest
```