[loader] Canonicalize structs by runtime ID + name
Rather than store repeated copies of a struct for each version of a
module (and each link context that that module is loaded in), keep a
single global cache of structs keyed on the struct type's fully
qualified name at runtime.

This is a suitable key because no matter which version of a package we
are running in, or which context it is loaded in, this will remain the
same. This ensures that the type loaded for a given Struct is
identical no matter the version it is loaded from or context it is
loaded in.

The changes yields some further simplifications:

- `resolve_struct_by_name` can use the new cache directly, and does
  not have to look for structs via a loaded module.
- ...this means that the loaded module no longer needs to keep track
  of a `struct_map`.
- ...this also means that it does not require access to the
  link_context.
- There is no need to have separate codepaths for `make_type` and
  `make_type_while_loading` because both implementations can use the
  `structs` cache

Test Plan:

Existing tests:

```
cargo nextest
```