Implement functions with multiple return values. (#105)
Functions in Move use a second-class tuple-like expression to bind, return, and destructure multiple values.

On exit from a function, we generate LLVM IR to wrap them up into a struct, which is returned as a single IR value. Similarly, when a callee that returns such a value is used in an expression, we generate IR to extract each actual value from the struct.

Also deduplicated load_call and load_call_store, as the former is just and instance the latter with no return values passed.

Added a move-ir-test and a runtime rbpf test to cover above.