Implement support for Move `signer`. (#157)
This patch adds support for the Move signer type. This is essentially
an opaque address (LLVM type `{ [N x i8] }' that can only be created or
accessed by the runtime.

Add missing native function signer::address_of(&signer): address to
the move-native library.

Add a new --signers command line option to move-mv-llvm-compiler which
allows the user or test harness to inject signer values to arguments of
type signer of the script function. This is for testing purposes only.
This is needed as there is no way for a user program to create a signer.

Implement support in the test harness for a new directive signers. This
allows rbpf tests to inject signer values for their unit tests. It parses
and feeds them to the --signers command line option.

Add a runnable test signer01 which shows all the pieces working together.