NFC: Replace many uplevel references to enclosing ModuleContext reference. (#153)
The FunctionContext has grown organically over time and now contains
an unwieldly collection of seven uplevel references to the enclosing
ModuleContext-- including boxed function pointers needed during the
function translation.

This patch replaces all the uplevels with a single reference to the enclosing
ModuleContext and updates all users. This will also simplify the passing of
module-level structures (llvm::Context, llvm::Module) that are commonly needed
in bundles, such as in the rttydesc module. A future patch will make those
sorts of simplifications.