[move-compiler v2] Porting remaining typing tests and refining type checker
This ports the remaining 200+ tests of the v1 compiler `typing` testsuite and refines the v2 compiler to handle those correctly. Major changes include:

- Introducing ordering in type unification. This is necessary to deal with conversions like `&mut T` to `&T`. The general mechanism will allow us to have also other conversions if we need down the road.
- Introducing a mechanism to define a default for type inference via constraints. If a type variable has no unique instantiation, then the default is chosen. For example, the default for integer constants is `u64`.
- Implementing the remaining expansion AST constructs from the Move parser. All expression variants are now handled.

With this, we should have a fairly feature complete and tested type checker. This doesn't do, however:

- visibility checking (private/friend)
- mutation checking (whether something which is mutated can be actually mutated)
- ability (copy, drop, ...) checking
- reference safety (borrow) checking

Those phases have been merged partly into the v1 type checker. For the v2 compiler, the plan is to handle them on bytecode level, similar as in the bytecode verifier.

GitOrigin-RevId: b69b4de30cd37f4ca6799c096b6c1069771ba3fd