[compiler v2] Bytecode generator
This implements and tests the stackless bytecode generator.

The stackless bytecode is the central intermediate representation for the compiler toolchain. It is generated from type checked AST. The initially generated bytecode is type-consistent but not yet checked for abilities, read-write sets/acquires, or reference safety. Also the bytecode is unoptimized. Those steps are expected to be implemented in later phases.

This generator should be feature complete.

GitOrigin-RevId: c322a403e131a07945c50e8bb1da8a576042a0fe