Extract entry function name before args passed in instruction_data
This is in preparation to passing all entry function parameters via
instruction_data byte array.

AccountInfo data must be used for reading from or writing to the
ledger. The only place that currently can be used to pass arbitrary
arguments is instruction_data. This change introduces a local variable
offset that is used to deserialize values of various types from the
instruction_data byte array. The offset is in an index in the byte
array pointing to the next argument to be deserialized. In this PR
offset is only used to extract the string name of the entry function
the entrypoint should pass control to. In the future the entry
function name may be replaced by a fixed-length hash code of the entry
function, and instead of comparing strings of characters the
entrypoint will compare the hash codes and call the corresponding
entry function.

The arguments of a specific entry function will follow the name string
of the entry function in the instruction_data byte array. The
generated entrypoint will deserialize the arguments from
instruction_data using the offset local variable. Address and signer
arguments will be matched with the elements of the accounts
AccountInfo vector. The total number of accounts in AccountInfo vector
should be equal to the total number of individual and bundled in
vectors of address or signer arguments of the entry function. The size
of the vector arguments will be serialized in the instruction_data
byte_array along with other argument values.