[compiler v2] Make type inference complete in the presence of structs (#11010)
* [compiler v2] Make type inference complete in the presence of structs

Closes #10884.

Until now, type inference had some ad-hoc treatment when it comes to structs. In the case of field access, or in the case of resource operations, the inferred type was forced to be concrete at the point of the occurence of the particular operation. For example, in `s.f`, if the type of `s` is not known at this program point, an error was produced. Similar, `move_from(a)` could not be used without explicit type instantiation because the inferred resource type isn't known yet.

In this PR we make "type inference whole" by removing the ad-hoc treatment. To this end, the type constraint system is extended by a new constraint `SomeStruct(fields)`. Thus for `s.f`, we can represent the type of `s` as 'some type X which is SomeStruct(f:Y)'. This brings us again a step closer to full support of traits in type inference.

* Addressing reviewer comments.

* Addressing reviewer comments.

* Fixing an issue with TryImplAsSpec mode

GitOrigin-RevId: 4a0702106f8a42f4fb5a91f1192ea26777e57ee7