[compiler v2] Relaxing copy restrictions on `&mut` (#11213)
Closes #11117.

This relaxes the restriction on copying of `&mut`. A copy of a `&mut` is now handled similar as a move, by creating a skip edge to the source. While that edge exists, the original source cannot be used, according to the existing borrowing rules.

Thus in a case like this:

```
let r = &mut <<some>>;
{
  let x = r;
  mutate(x);
}
mutate(r);
```

The 'copy' of x to r is valid.  As long as the inner `x` lives, the outer `r` cannot be used. When `x` goes out of scope, `r` can be used again. Thus, while copies of mut-ref values is now possible, only one 'active' can exist at a time.

This PR also does the following:

- fixes a bug in reference-safety, resulting in more errors. v1 does not have those errors, but they appear justified. This is another gray zone in borrow semantics, so I think those errors should stay until we have counter examples in real code.
- enables reference-safety (it wasn't on in regular execution, only in tests)
- activates experiments and adds a first one `--experiment=no-safety` (or `MOVE_COMPILER_EXP=no-safety`) which allows to skip safety checks (currently visibility and reference safety). The env var should work if called via the aptos CLI or unit test framework.

GitOrigin-RevId: 666512b211b20eab94202938c371198ec56300cb