[compiler-v2] Only aggressively copy if type has drop ability (#11510)
* [compiler-v2] Only aggressively copy if type has drop ability

The hack to work around #11223 in live var analysis may cause ability errors if the value does not have the drop ability. This is because if we copy to aggressively, we may need to destroy values later, and this requires those to have a drop ability. This change fixes this by requiring to have drop to apply the hack.

* Enable the fix, demonstrating the delta

GitOrigin-RevId: 66d35fb770cf8908b4e79192c5a4a0cda063a7c2