[compiler-v2] Fixes a bug in file-format generator and a bug in bytecode generator (#11519)
* If a value is pushed on the stack from a call and is used more than once after that, the stack must be flushed
so the value is saved in register, as illustrated by this example.

Illustrates that values on the stack must be flushed if used again after a call.

* [compiler-v2] Flush the abstract stack in FF generator if values are still used

If a value is left on the stack from a previous call and then passed on to a subsequent call, the abstract stack in `function_generator.rs` tries to avoid to store it in a local but instead to directly consume from the stack. However, if that value is used afterwards again, it needs to be poped from the stack and stored in a local. In this PR `abstract_push_args` is modified to achieve this.

The 1st commit demonstrates the problem in test case `flush_if_alive.move`. The 2nd commit fixes the issue.

* Demonstrates a missing Deref on the operand of `borrow_global<ObjectCore>(o.inner)`. If `o.inner` is stored in a `let` the problem isn't there.

* [compiler-v2] Do not propagate reference mode to nested expressions in bytecode generator

For a nested field selection as in `(x.f).g`, the inner field selection must be compiled in reference mode, creating a reference for `x.f` and for `x`.  However, the reference mode must not be applied to nested expressions. For example, in `borrow_global<T>(x.foo).bar`, `x.foo` must not be compiled in reference mode. This PR introdues a way to escape reference mode for nested expressions and applies it where appropriate in the bytecode generator.

The 1st commit demonstrates the bug, the 2nd fixes it. The test case file has been renamed to match the actual problem being fixed as it became clearer during debugging.

GitOrigin-RevId: c44eb495c0c8c20e74fc7c1fa6b24924a9f09295