[compiler-v2] Lambda lifting + first steps towards a pipeline for ast trafo (#12318)
* [compiler-v2] Lambda lifting + first steps towards a pipeline for ast trafo

This implements lambda lifting as rewrite on `GlobalEnv`. Lambda lifting appeared necessary for prover connection, because this is how the v1 integration works. Its not that clear any longer whether lambda lifting is really needed, but now the work is done, and we shoukd preseve it for future general support of HOFs.

As this is now the 5th or so so transformation on the environment before the bytecode pipeline starts. to support this adequate in the test suite, there is also a first step towards an `EnvProcessorPipeline`, and a new submodule `env-pipeline` which should contain those processors. For now it is just lambda lifting.

* Addressing reviewer comments

* Addressing reviewer comments

* Addressing reviewer comments

GitOrigin-RevId: e3cdf2041676c70f4da081c53c102b7430bd2ca5