[move-compiler-v2] Implement AST Simplification to solve #10850 (#11553)
This is an expression simplifier which works at the AST level in Compiler V2,
doing some quick optimizations that reduce the amount of code the stackless
bytecode passes have to deal with. Notably, it eliminates some of the overhead
of inlining, as seen in the test case deep_exp.move.

* Fixes #10850: Compiler V2 needs expression simplification
* Fixes #11535: Constant folding should not allow full-bitsize shifts
* Fixes #11536: Constant folding displays redundant errors
* Fixes #11549: Model Dump for a Cast operation doesn't show the type
* Fixes #11805: Load constants the right way
* Fixes #11720: Process got killed when running unit tests in aptos-gas-schedule-updator using V2

GitOrigin-RevId: d6acef0e46e6d5c036bb7ebd669083a74ab9caf4