[compiler-v2] Introduces custom rules for checking the freeze operation (#12725)
* [compiler-v2] Introduces custom rules for checking the freeze operation

The freeze operation has some unexpected ad-hoc rules in the v1 borrow semantics: when a reference is frozen, there must not exist any other mutable reference pointing to the same location. It is, however, ok if the currently frozen reference is used later again. This seems to be over-restrictive since it shouldn't matter whether copies of the mutable reference exist as long as they aren't passed at the same time as an argument to a function, i.e. as long as they do not create aliasing.

This PR simulates the v1 borrow rules for freeze, so we avoid bytcode verification errors. At later point we may relax the rules in the bytecode verifier, at which point we could remove those rules or guard them behind a flag.

Closes #12554.  Als repairs some of the bugs in #12701, but not all, so keeping this bug open.

* Adding new test case

GitOrigin-RevId: 5f9906f2481ff4a0d31ebbf06840fa9bac7bd555